requirements.txt
django==3.2.6
coreapi
Pillow